changes :-
- restructed code base
- added back queuing system metioned in #14
- removed zipping if the mediagroup is above items are above 10
- temporaryly removed logging
- other minor changes
