+++ __main__.py
@@ -1,0 +1,7 @@
+from webshotbot import WebshotBot
+import pyrogram
+
+if __name__ == "__main__":
+    bot = WebshotBot()
+    bot.start()
+    pyrogram.idle()
