+++ engine/browser.py
@@ -1,0 +1,77 @@
+# (c) AlenPaulVarghese
+# -*- coding: utf-8 -*-
+
+from http.client import BadStatusLine, ResponseNotReady
+from helper.images import draw_statics
+from pyppeteer.browser import Browser
+from pyppeteer import launch, errors
+from helper import Printer
+import logging
+import asyncio
+import os
+
+_LOG = logging.getLogger(__name__)
+_EXEC_PATH = os.environ.get("GOOGLE_CHROME_SHIM", None)
+
+
+async def launch_browser(retry=False) -> Browser:
+    try:
+        _LOG.info("Launching browser")
+        browser = await launch(
+            headless=True,
+            logLevel=50,
+            executablePath=_EXEC_PATH,
+            handleSIGINT=False,
+            args=[
+                "--no-sandbox",
+                "--single-process",
+                "--disable-dev-shm-usage",
+                "--disable-gpu",
+                "--no-zygote",
+            ],
+        )
+        return browser
+    except BadStatusLine as e:
+        if not retry:
+            _LOG.error("Launching browser failed due to %s, retrying...", e)
+            await asyncio.sleep(1.5)
+            return await launch_browser(True)
+        elif retry:
+            _LOG.error("Launching browser failed due to %s, exiting...", e)
+
+
+async def screenshot_engine(browser: Browser, printer: Printer):
+    page = await browser.newPage()
+    await page.setViewport(printer.resolution)
+    try:
+        await page.goto(printer.link)
+        title = await page.title()
+        printer.slugify(title[:14])
+        if printer.type == "pdf":
+            await page.pdf(printer.arguments_to_print, path=printer.file)
+        elif printer.type == "statics":
+            height, width, metrics = await asyncio.gather(
+                page.evaluate(
+                    "Math.max(document.body.scrollHeight, document.body.offsetHeight, "
+                    "document.documentElement.clientHeight, document.documentElement.scrollHeight, "
+                    "document.documentElement.offsetHeight);"
+                ),
+                page.evaluate(
+                    "Math.max(document.body.scrollWidth, document.body.offsetWidth, "
+                    "document.documentElement.clientWidth, document.documentElement.scrollWidth, "
+                    "document.documentElement.offsetWidth);"
+                ),
+                page.metrics(),
+            )
+            page_data = dict(Height=height, Width=width)
+            page_data.update(metrics)
+            byteio_file = await asyncio.get_running_loop().run_in_executor(
+                None, draw_statics, title[:25], page_data
+            )
+            printer.set_location(byteio_file)
+        else:
+            await page.screenshot(printer.arguments_to_print, path=printer.file)
+    except errors.PageError:
+        raise ResponseNotReady("This is not a valid link 🤔")
+    finally:
+        await page.close()
