--- plugins/logger.py
@@ -1,9 +1,0 @@
-import logging
-
-
-logging.basicConfig(
-    level="WARNING",
-    format="%(asctime)s - %(name)s - [ %(message)s ]",
-    datefmt="%d-%b-%y %H:%M:%S",
-    handlers=[logging.FileHandler("debug.log"), logging.StreamHandler()],
-)
